restructured for proegarting update of points. No SQL in service layer

+++ backend/service/adminService.js
@@ -1,4 +1,6 @@
 const adminRepo = require('../repository/adminRepository');
+const ctRepo = require('../repository/ctRepoHelper');
+const subsRepo = require('../repository/subsRepoHelper');
 
 function createContest({ contestName, maxSubs }) {
     return adminRepo.createContest({ contestName, maxSubs });
@@ -8,11 +10,22 @@ function createUser({ username, password, role }) {
     return adminRepo.createUser({ username, password, role });
 }
 
+function addToContest({ contestID, teamID }) {
+    return adminRepo.addToContest(contestID, teamID);
+}
+
 function changeMaxSubs(contestID, newMaxSubs) {
     return adminRepo.changeMaxSubs(contestID, newMaxSubs);
 }
 
-function deleteSubmission(submissionID) {
+async function deleteSubmission(submissionID) {
+    const submission = await subsRepo.getSubmission(submissionID);
+
+    if (!submission) {
+        throw new Error('Submission not found');
+    }
+
+    await ctRepo.removePoints(submissionID, submission.teamID, submission.contestID);
     return adminRepo.deleteSubmission(submissionID);
 }
 
@@ -30,5 +43,6 @@ module.exports = {
     changeMaxSubs,
     deleteSubmission,
     makeActive,
-    makeInactive
+    makeInactive,
+    addToContest
 };
